Hotfix minor torch scan

+++ src/scan_single_core.cpp
@@ -31,10 +31,10 @@ extern "C" __global__ __aicore__ void scan_single_core_int8(GM_ADDR vec_in,
   const uint32_t matmul_size = tiling.matmul_size;
 
   GM_ADDR const usrWorkspace = AscendC::GetUserWorkspace(workspace);
-  GM_ADDR const lower = load_tril_matrix<half>(matmul_size);
+  GM_ADDR const lower = load_tril_matrix<int8_t>(matmul_size);
 
-  run_scan_single_core<half>(vec_in, lower, vec_out, vec_len, matmul_size,
-                             usrWorkspace);
+  run_scan_single_core<int8_t>(vec_in, lower, vec_out, vec_len, matmul_size,
+                               usrWorkspace);
 }
 
 /**
